+++ cint/cint/src/global.h
@@ -279,7 +279,7 @@ extern int G__eof_count;        /* end of file error flag */
 extern int G__ismain;           /* is there a main function */
 extern int G__globalcomp;       /* make compiled func's global table */
 extern int G__store_globalcomp;
-extern long G__globalvarpointer; /* make compiled func's global table */
+extern size_t G__globalvarpointer; /* make compiled func's global table */
 extern int G__isfilebundled;       /* indicate if we have a bundle file in rootcint */
 
 #ifdef __cplusplus
